HOTFIX: improve support for WPML
This should be refactored. Unnecessary repetition.

If WPML is active, than the size of the terms will be depending on the most and least popular term IN ANY LANGUAGE.
So the check if a term is more popular than the most popular so far has to happen after the decision if a term in included in the tag cloud.